+++ include/readseek_impl.h
@@ -16,8 +16,8 @@ public:
 
   virtual uint64_t getID() const override { return 0; }
 
-  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
-  virtual int64_t read(size_t len, uint8_t* buf) override;
+  virtual size_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual size_t read(size_t len, uint8_t* buf) override;
 
   virtual size_t tellg() const override { return Pos; }
   virtual size_t seek(size_t pos) override { return (Pos = (pos < Buf.size() ? pos: Buf.size())); }
@@ -42,8 +42,8 @@ public:
 
   virtual uint64_t getID() const override { return 0; }
 
-  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
-  virtual int64_t read(size_t len, uint8_t* buf) override;
+  virtual size_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual size_t read(size_t len, uint8_t* buf) override;
 
   virtual size_t tellg() const override;
   virtual size_t seek(size_t pos) override;
@@ -70,8 +70,8 @@ public:
 
   virtual uint64_t getID() const override { return Inum; }
 
-  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
-  virtual int64_t read(size_t len, uint8_t* buf) override;
+  virtual size_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual size_t read(size_t len, uint8_t* buf) override;
 
   virtual size_t tellg() const override { return 0; }
   virtual size_t seek(size_t pos) override;
